requirements.txt
openai==2.6.1
langchain==1.1.0
langchain-openai==1.1.0
python-dotenv==1.2.1
env_utils==1.2.0
langfuse==3.10.5
streamlit==1.52.0
pypdf==6.4.0
langchain-chroma==1.0.0
chromadb==1.3.5
langchain_community==0.4.1
mlflow==3.6.0
databricks-langchain==0.11.0
langchain-ollama==1.0.0
